Ontology caching and significant restructuring.
The HelloWorld class (which should be renamed to Turbulence) now spawns threads
for TurbulenceDriver.

TurbulenceDriver manages all ontology and clusterspace classes. It uses a queue
to listen to incoming requests and processes them. Currently all requests are
implicit registerSchema requests, which will be refined later.

The OntologyMapper attempts to load a cached ontology by calculating the SHA1
of the canonical URI and fetching from the ontologies directory.  If it does
not exist, the driver will call OntologySaver to create it.

Turbulence has finally started using the configuration file.