Index known ontologies & track sources
Known ontologies are indexed to prevent duplicates. Each one should be rooted
to the special reference node rather than to the graph reference node.

Each class maintains a link to this so we can track individual data sets when
they are requested.

The whole link operation is now wrapped in one transaction so that it is all or
nothing.