ClusterSpaceJenaGraph: Jena <-> ClusterSpace bridge
Since we use Jena ARQ for SPARQL queries, we need a way for Jena to treat the
clusterspace the way Jena represents graphs internally. Then we also need to
process triple queries on the clusterspace based on our interpretations of
certain relationships, and mapping other relationships to neo4j relationships.